refactor: replace searchbar condition to constant
