+++ .env
@@ -1,1 +1,1 @@
-API_KEY=AIzaSyADhA_6GT3-_EI4BKMZsGBFMQdClqXuyHA
+YOUTUBE_API_KEY=AIzaSyB9KMr47y-jQHDR7beRpnyPbQgKoof2Vho
