+++ src/constants/api.ts
@@ -1,5 +1,5 @@
 const SEARCH_URL = 'https://youtube.googleapis.com/youtube/v3/search';
 const VIDEO_DATA_URL = 'https://www.googleapis.com/youtube/v3/videos';
-const KEY = process.env.API_KEY;
+const KEY = process.env.YOUTUBE_API_KEY;
 
 export { KEY, SEARCH_URL, VIDEO_DATA_URL };
